Relocate tests in submodules
